refac: move password inputs into component
